drivers/virtualbox: encapsulate host-only IP resolution in driver
